Modified requirement 7
I have clarified the wording of the requirement to ensure that it can be
understood that *usability*, and not *performance* is the area being
covered here. I also reduced the specified amount of time after another
comment.

A further change was made in the wording to clear up some slightly
ambigious language.